tools/global.cpp: assign default verbose value.
After commit 3c4e802638d4 ("Completely Rewrite Command-Line Argument
Parsing"), the "verbose" option was assigned an implicit value (i.e.
the value used for "-v" without an explicit number), but not a default
value (i.e. the value used if no "-v" is given). Furthermore, if the
default value "0" is given, no change to the runtime setting is made

This created two problems:

1. If openEMS is used as a shared library in Python, and the first
simulation call is made with a non-default verbose value, but the
second call is made with a default value, the second call has no
effect - because the default value is only applied once on startup.

This commit assigns a default value of 0 for "verbose" and applies
this value on every invocation of the argument parser.

2. The help message doesn't show the default value "0", only the
implicit value "1":

    -v [ --verbose ] [=arg(=1)]  Verbose level, select debug level 1 to 3, also
                                 accept -v, -vv, -vvv

Furthermore, the help message now correct shows both the default
and implicit value for "verbose":

    -v [ --verbose ] [=arg(=1)] (=0)  Verbose level, select debug level 1 to 3,
                                      also accept -v, -vv, -vvv

Signed-off-by: Yifeng Li <tomli@tomli.me>