tools/tiling: FDTD temporial tiling algorithm implementation.
FDTD simulation is a form of stencil computation, its order
of calculations can be changed in such a way to partition
the simulation domain into multiple small regions, in which
multiple timesteps can be calculated at a time. It increases
simulation speed significantly due to better cache locality.
This technique is known as tiling or blocking. More
specifically, temporal tiling.

Correct tiling is tricky to implement and test, so the plan,
thus, tiling.h is a library for pre-calculate the necessary
coordinates of all regions according to the parallelogram and
trapezoid tiling algorithms.

This is used by the Tiling engine to generate a "plan" before
starting the simulation.

For more information, see:

* Fukaya, T., & Iwashita, T. (2018). Time-space tiling with
  tile-level parallelism for the 3D FDTD method. Proceedings of
  the International Conference on High Performance Computing
  in Asia-Pacific Region - HPC Asia 2018.

  doi: 10.1145/3149457.3149478

* The original implementation of this library, with notes and
  diagnostic tools: https://github.com/biergaizi/project-diamond

Signed-off-by: Yifeng Li <tomli@tomli.me>