Engine_Extension: add Tiling API
Using the existing engine extension APIs, the internal states
of the engine and the voltages and currents of the simulation
domain can be updated in two ways:

1. Updating the entire simulation domain by not passing additional
argument.

2. Asking the extensions to decide where to update by themselves,
based on its "int threadID".

To allow the new engine implementation with spatial and temporal
tiling, this commit introduces a new set of engine extention APIs
for use specifically by the Tiling engine:

    virtual void InitializeTiling(const Tiling::Plan3D& plan);

    DoPreVoltageUpdates(int timestep, Tiling::Range3D<> range);
    DoPostVoltageUpdates(int timestep, Tiling::Range3D<> range);
    Apply2Voltages(int timestep, Tiling::Range3D<> range);
    DoPreCurrentUpdates(int timestep, Tiling::Range3D<> range);
    DoPostCurrentUpdates(int timestep, Tiling::Range3D<> range);
    Apply2Current(int timestep, Tiling::Range3D<> range);

In these APIs, the updated range is strictly contained within Tiling::Range3D,
which contains start and stop coordinates in 3D space. It's the engine's
responsibility to partition the space into tiles, not the extentions.
Furthermore, since the engine may calculate many timesteps at once in the
same tiles, but some extentions may do something only at a specific timestep,
the tiling API also needs to know the current timestep.

To implement tiling in an extention, the engine should check whether the
current range is relevant, if so, this range is updated. If not, this
range is ignored. The updated range should not depend on the values from
other ranges - "action at a distance" is not allowed.

Further, the InitializeTiling() method is used for passing all tiles
involved to the extentions in advance for initialization, it's only
used by the tiling engine, in analogues to the role of the
SetNumberOfThreads() method used by the standard multi-threading
engine. That is, the space partitioning plan is sent to each extension
so that they can have prior information about which tiles need processing,
and which tiles do not. In the future, this step may also be used for
zeroing memory for NUMA first touch.

Signed-off-by: Yifeng Li <tomli@tomli.me>