tools/arraylib/simd.h: new SIMD wrapper class
Class simd.h is a new wrapper for manipulating SIMD/SSE vector
data. In comparison to the original f4vector class, it has two
advantages:

1. It's templated so that one can create a vector of arbitrary data
type and length, such as:

    Simd<int, 4> v1, v2;
    Sime<float, 8> v3;

2. Individual elements can be accessed by the natural operator[]
syntax, such as:

    v1[2]
    v2[7]

3. The underlying union is private, operations are applied directly
on a vector via overloaded operators (without the member variable
".v"), such as:

    v1 += v2;

4. Implement operator== using memcmp, so manual vector comparison
is not necessary.

5. It relies on vector extension in GCC/clang. Manual hardcoded
fallbacks for each data size and type are needed for MSVC. Only
float4 is implemnted for now. Since clang can be used as the MSVC
complier, it's probably a good idea to make the switch?

Signed-off-by: Yifeng Li <tomli@tomli.me>