CMakeLists.txt: drop HDF5 version check.
HDF5's CMake rules have a bug when running on FreeBSD,
making it to believe HDF5 1.10 is older than HDF5 1.8 [1],
thus CMake fails. Furthermore, implementing a manual
version check is also problematic since it also has a bug
on Linux so it doesn't define HDF5_VERSION_MAJOR, worse,
HDF5_VERSION is only defined in CMake 3.3 so it won't work
on old systems that actually need this check. To simplify
the workaround we simply drop the HDF5 version check (since
HDF 1.8 was released in 2008, while HDF 1.6 was last updated
in 2011, it's unlikely to cause any issue).

[1] https://gitlab.kitware.com/cmake/cmake/-/issues/25358

Signed-off-by: Yifeng Li <tomli@tomli.me>